Get the last 100 open tickets for pagure, gitlab and github.
This commit refactors how we fetch the tickets from pagure, github and gitlab.
Instead of getting all tickets we care only for the Open tickets. So far we do
not manage pagination and we only fetch the last 100 open tickets which is in most case
all of the tickets.
Also this commit removes the logic to deal with the closed tickets.

Fixes #16

Signed-off-by: Clement Verna <cverna@tutanota.com>